requirements.txt
opencv-python
numpy
flask
pillow
gunicorn